dom/tests/mochitest/gamepad/test_navigator_gamepads.html

<!-- Any copyright is dedicated to the Public Domain.
   - http://creativecommons.org/publicdomain/zero/1.0/ -->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test gamepad</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script type="text/javascript" src="mock_gamepad.js"></script>
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();

var testNum = 0;
var tests = [
  check_first_gamepad,
  check_second_gamepad,
  check_gamepad_hole,
  check_no_gamepads,
];

function run_next_test(event) {
  SpecialPowers.executeSoon(function() { tests[testNum++](event); });
}

function buttonhandler(e) {
 run_next_test(e);
}

function disconnecthandler(e) {
  run_next_test(e);
}
window.addEventListener("gamepadbuttondown", buttonhandler);
window.addEventListener("gamepaddisconnected", disconnecthandler);

runGamepadTest(startTest)

function startTest() {
  // gamepads should be empty first
  is(navigator.getGamepads().length, 0, "should be zero gamepads exposed");
  // Add a gamepad
  GamepadService.addGamepad("test gamepad 1", // id
                     GamepadService.standardMapping,
                     4, // buttons
                     2).then(function(index) {
                       internal_index1 = index;
                       // Press a button to make the gamepad visible to the page.
                       GamepadService.newButtonEvent(internal_index1, 0, true);
                     });
}

var content_index1 = 0;
var internal_index2;
var content_index2 = 1;

function check_first_gamepad(e) {
  ok(true, "Checking first gamepad");
  // First gamepad gets added.
  is(e.gamepad.id, "test gamepad 1", "correct gamepad name");
  var gamepads = navigator.getGamepads();
  is(gamepads.length, 1, "should have one gamepad exposed");
  is(gamepads[e.gamepad.index], e.gamepad, "right gamepad exposed at index");
  is(gamepads[content_index1], e.gamepad, "gamepad counter working correctly");
  // Add a second gamepad, should automatically show up.
  GamepadService.addGamepad("test gamepad 2", // id
                     GamepadService.standardMapping,
                     4, // buttons
                     2).then(function(index) {
                       internal_index2 = index;
                       GamepadService.newButtonEvent(internal_index2, 0, true);
                     });
  ok(true, "Done checking first gamepad");
}

function check_second_gamepad(e) {
  ok(true, "Checking second gamepad");
  // Second gamepad gets added.
  is(e.gamepad.index, 1, "gamepad index should be 1")
  is(e.gamepad.id, "test gamepad 2", "correct gamepad name");
  var gamepads = navigator.getGamepads();
  is(gamepads.length, 2, "should have two gamepads exposed");
  is(gamepads[e.gamepad.index], e.gamepad, "right gamepad exposed at index");
  is(gamepads[content_index2], e.gamepad, "gamepad counter working correctly");
  // Now remove the first one.
  GamepadService.removeGamepad(internal_index1);
  ok(true, "Done checking second gamepad");
}

function check_gamepad_hole(e) {
  ok(true, "Checking gamepad hole");
  // First gamepad gets removed.
  var gamepads = navigator.getGamepads();
  is(gamepads.length, 2, "gamepads should have two entries");
  is(gamepads[content_index1], null, "should be a hole in the gamepad list");
  isnot(gamepads[content_index2], null, "second gamepad should exist");
  // Now remove the second one.
  GamepadService.removeGamepad(internal_index2);
  ok(true, "Done checking gamepad hole");
}

function check_no_gamepads(e) {
  ok(true, "Checking no gamepads");
  // Second gamepad gets removed.
  var gamepads = navigator.getGamepads();
  is(gamepads.length, 0, "gamepads should be empty");
  SimpleTest.finish();
}
</script>
</body>
</html>
